Create StudentRegistrationCode
This code resembles the process of student registration with major details provided in the task. this code consists of two functions namely add student and generate report. add student function get the details from the user and save it to file for multiple students and generate report function provide the details of total fees of students.